Rename tagletJdk11 -> taglet, taglet -> tagletJdk8
