+++ framework-test/build.gradle
@@ -2,7 +2,7 @@ import org.gradle.internal.jvm.Jvm
 
 sourceSets {
     taglet
-    tagletJdk11
+    tagletJdk8
 }
 
 dependencies {
@@ -13,7 +13,7 @@ dependencies {
     implementation "org.plumelib:plume-util:${plumeUtilVersion}"
 
     if (Jvm.current().toolsJar) {
-        tagletImplementation files(Jvm.current().toolsJar)
+        tagletJdk8Implementation files(Jvm.current().toolsJar)
     }
 }
