+++ framework-test/src/taglet/java/org/checkerframework/taglet/ManualTaglet.java
@@ -1,12 +1,19 @@
 // Keep this file in sync with
-// ../../../../../tagletJdk11/java/org/checkerframework/taglet/ManualTaglet.java .
+// ../../../../../tagletJdk8/java/org/checkerframework/taglet/ManualTaglet.java .
 
 package org.checkerframework.taglet;
 
-import com.sun.javadoc.Tag;
-import com.sun.tools.doclets.Taglet;
-import java.util.Map;
+import com.sun.source.doctree.DocTree;
+import com.sun.source.doctree.TextTree;
+import com.sun.source.doctree.UnknownBlockTagTree;
+import com.sun.source.doctree.UnknownInlineTagTree;
+import com.sun.source.util.SimpleDocTreeVisitor;
+import java.util.EnumSet;
+import java.util.List;
+import java.util.Set;
 import java.util.StringJoiner;
+import javax.lang.model.element.Element;
+import jdk.javadoc.doclet.Taglet;
 
 /**
  * A taglet for processing the {@code @checker_framework.manual} javadoc block tag, which inserts
@@ -23,39 +30,18 @@ import java.util.StringJoiner;
  */
 public class ManualTaglet implements Taglet {
 
+  private static final String NAME = "checker_framework.manual";
+
   @Override
   public String getName() {
-    return "checker_framework.manual";
+    return NAME;
   }
 
-  @Override
-  public boolean inConstructor() {
-    return true;
-  }
+  private final EnumSet<Location> allowedSet = EnumSet.allOf(Location.class);
 
   @Override
-  public boolean inField() {
-    return true;
-  }
-
-  @Override
-  public boolean inMethod() {
-    return true;
-  }
-
-  @Override
-  public boolean inOverview() {
-    return true;
-  }
-
-  @Override
-  public boolean inPackage() {
-    return true;
-  }
-
-  @Override
-  public boolean inType() {
-    return true;
+  public Set<Taglet.Location> getAllowedLocations() {
+    return allowedSet;
   }
 
   @Override
@@ -94,32 +80,46 @@ public class ManualTaglet implements Taglet {
   }
 
   @Override
-  public String toString(Tag tag) {
-    String[] split = tag.text().split(" ", 2);
-    return formatHeader(formatLink(split));
-  }
-
-  @Override
-  public String toString(Tag[] tags) {
-    if (tags.length == 0) {
+  public String toString(List<? extends DocTree> tags, Element element) {
+    if (tags.isEmpty()) {
       return "";
     }
     StringJoiner sb = new StringJoiner(", ");
-    for (Tag t : tags) {
-      String text = t.text();
+    for (DocTree t : tags) {
+      String text = getText(t);
       String[] split = text.split(" ", 2);
       sb.add(formatLink(split));
     }
     return formatHeader(sb.toString());
   }
 
-  @SuppressWarnings({"unchecked", "rawtypes"})
-  public static void register(Map tagletMap) {
-    ManualTaglet tag = new ManualTaglet();
-    Taglet t = (Taglet) tagletMap.get(tag.getName());
-    if (t != null) {
-      tagletMap.remove(tag.getName());
-    }
-    tagletMap.put(tag.getName(), tag);
+  static String getText(DocTree dt) {
+    return new SimpleDocTreeVisitor<String, Void>() {
+      @Override
+      public String visitUnknownBlockTag(UnknownBlockTagTree node, Void p) {
+        for (DocTree dt : node.getContent()) {
+          return dt.accept(this, null);
+        }
+        return "";
+      }
+
+      @Override
+      public String visitUnknownInlineTag(UnknownInlineTagTree node, Void p) {
+        for (DocTree dt : node.getContent()) {
+          return dt.accept(this, null);
+        }
+        return "";
+      }
+
+      @Override
+      public String visitText(TextTree node, Void p) {
+        return node.getBody();
+      }
+
+      @Override
+      protected String defaultAction(DocTree node, Void p) {
+        return "";
+      }
+    }.visit(dt, null);
   }
 }
